updated the test drivers to deallocate the memory and changed dim type to int

+++ elina_poly/tests/libFuzzer/test_poly.h
@@ -13,14 +13,14 @@
 #define R(i) i < K
 
 bool create_polyhedron(opt_pk_array_t** polyhedron, elina_manager_t* man, opt_pk_array_t * top,
-		long dim, const long *data, size_t dataSize,
+		int dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp);
 
 bool make_fuzzable(void *array, size_t size, const long *data, size_t dataSize,
 		unsigned int *dataIndex);
 bool assume_fuzzable(bool condition);
 
-bool make_fuzzable_dimension(long * dim, const long *data, size_t dataSize,
+bool make_fuzzable_dimension(int * dim, const long *data, size_t dataSize,
 		unsigned int *dataIndex, FILE *fp);
 
 #endif /* TEST_POLY_H_ */
